Refactor, created separate YoutubeVideo component
