+++ src/shared/ui/survey/index.ts
@@ -4,5 +4,6 @@ export { StackedRadiosItem } from './StackedRadiosItem';
 export { StackedCheckBoxItem } from './StackedCheckboxItem';
 export { default as AudioPlayer } from './AudioPlayer';
 export { default as VideoPlayer } from './VideoPlayer';
+export { default as YoutubeVideo } from './YoutubeVideo';
 export * from './Slider';
 export * from './CheckBox';
